be/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
cryptography==41.0.5
bcrypt==3.2.2
python-multipart==0.0.6
jinja2==3.1.2
langchain-core==0.1.4
neo4j
python-jose[cryptography]==3.3.0
alembic==1.12.1
mysql-connector-python
pymysql
py_vncorenlp
google-generativeai
fastembed
torch
torchvision
torchaudio
torch-geometric
sentence_transformers
qdrant-client
langchain-nvidia-ai-endpoints
email-validator
passlib[bcrypt]==1.7.4
python-jose
boto3
openai
pdfplumber
langchain-nvidia-ai-endpoints
cohere



